auth/build_image/requirements.txt
flask
flask-bcrypt
flask-views
pyjwt
